backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.5
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt,argon2]>=1.7
python-multipart>=0.0.6
python-dotenv>=1.0
cryptography>=42.0
twilio>=9.0
boto3>=1.34
